+++ db/migrate/20150310181807_create_questions.rb
@@ -1,0 +1,11 @@
+class CreateQuestions < ActiveRecord::Migration
+  def change
+    create_table :questions do |t|
+      t.integer :challenge_id
+      t.string :name
+      t.boolean :deleted, null: false, default: false
+
+      t.timestamps null: false
+    end
+  end
+end