+++ db/migrate/20150310181912_create_answers.rb
@@ -1,0 +1,12 @@
+class CreateAnswers < ActiveRecord::Migration
+  def change
+    create_table :answers do |t|
+      t.integer :challenge_id
+      t.integer :question_id
+      t.integer :user_id
+      t.text :response
+
+      t.timestamps null: false
+    end
+  end
+end